Refactor into modules, fix bugs, expand tests, rewrite README
- Split monolithic src/typectl.ts into focused modules: types.ts,
  polyfill.ts, wrap.ts, pick.ts, concurrency.ts, iterate.ts,
  mappers.ts, assign.ts; typectl.ts is now a pure barrel re-export
- Fix for...in → Object.entries() in iterate (avoids inherited keys)
- Fix recursive ReadableStream pump → iterative while loop (no stack risk)
- Fix pick error test: silent try/catch → expect().rejects.toThrow()
- Add JSDoc to every exported function covering behaviour, concurrency
  semantics, and examples
- Add 11 edge-case tests: empty array/record/stream for iterate, toArray,
  toRecord, toValue, toStream; wrap with zero args; all/each on []
- Fix package.json exports: condition-specific "types" paths for ESM and
  CJS consumers; add "files": ["dist"] whitelist; remove jsx lint-staged glob
- Fix bin/watch to watch both CJS and ESM targets in parallel
- Fix README stale spec example (Jest import → Vitest)
- Rewrite README with motivation, full API reference with examples,
  concurrency table, type export table, compatibility, and dev commands

Made-with: Cursor
